fixed bugs
General bug fixing and compensation for data type of input depth map

+++ research/echos/circleFilter.m
@@ -8,7 +8,7 @@ r = round(r/x);
 
 cx=floor(width/2);cy=floor(len/2);ix=width;iy=len;
 [x,y] = meshgrid(-(cx-1):(ix-cx),-(cy-1):(iy-cy));
-cmask =((x.^2+y.^2)<=r^2);
+cmask = double(((x.^2+y.^2)<=r^2));
 %imagesc(d2)
 d2 = cmask.*d;
 imagesc(d2);
